refactor: 프로젝트 구조 개선 (main 파일을 components, layouts, pages로 분리)
